build: do not run PHPUnit from `composer test`
On CI (Quibble) we invoke `composer test` immediately after the
repositories have been closed by before MediaWiki is installed. Since
`tests/phpunit/phpunit.php` includes `Setup.php` which in turn tries to
include `LocalSettings.php`, that fails on CI.

Align `composer test` with the BoilerPlate extension composer.json which
is to only run linters. The PHPUnit tests are run in a different job in
CI.

The `composer phpunit` script is left as a convenience.

Bug: T404296
Change-Id: Ic0c123ffd6e9675a13e685812fdebe3db4fd2d19