Add option to pass additional parameters to geolocation API
This commit adds a configuration parameter, `$wgCookieConsentGeolocationParameters`, that allows an administrator of a wiki to pass additional query parameters to the geolocation API. This makes it possible to e.g. pass an API key. It also changes the default API from the free version of ip-api.com to the premium version in order to prevent accidental commercial usage.

Change-Id: I85a4abaa92a9db0e866b792f9662d97a59f33a48